backend/morec/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
python-dotenv==1.0.0
Pillow==10.0.0
drf-yasg==1.21.6
djoser==2.2.0
django-filter==23.2
gunicorn==21.2.0
psycopg2-binary==2.9.6
coreapi==2.3.3
django-cors-headers==4.2.0
PyJWT==2.7.0
celery==5.3.1
redis==4.6.0
django-import-export==3.2.0
social-auth-app-django==5.3.0

# pandas==2.1.1
# scikit-learn==1.1.3
# tensorflow==2.13.0
# dill==0.3.6
